add simple sftp ingestion for initial ringgo testing
