--- lambdas/sftp-to-s3/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "liberator-sftp-to-s3",
-  "version": "1.0.0",
-  "description": "A function to move liberator zip files from an sftp server to an s3 bucket",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "aws-sdk": "^2.913.0",
-    "ssh2-sftp-client": "^6.0.1"
-  }
-}
