--- terraform/modules/sftp-to-s3/02-inputs-optional.tf
@@ -1,29 +1,0 @@
-variable "lambda_execution_cron_schedule" {
-  description = "CRON expression to schedule the Lambda"
-  type        = string
-  default     = "cron(0 6 * * ? *)"
-}
-
-variable "ephemeral_storage" {
-  description = "Amount of temporary storage in MBs"
-  type        = number
-  default     = 512
-}
-
-variable "lambda_timeout" {
-  description = "Lambda time out in seconds"
-  type        = number
-  default     = 900
-}
-
-variable "lambda_memory_size" {
-  description = "Memory for lambda in MBs"
-  type        = number
-  default     = 128
-}
-
-variable "glue_job_to_trigger" {
-  description = "Name of Glue job to trigger once data has been ingested to S3"
-  type        = string
-  default     = ""
-}
